requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-shopping-cart==0.1
fontawesomefree==6.1.2
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
mc-bin-client==1.0.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dateutil==1.5
requests==2.28.1
sqlparse==0.4.2
stripe==4.1.0
tap==0.2
urllib3==1.26.12
whitenoise==6.2.0
